first api call displays the popular movies

+++ app/src/main/java/com/moviedb/practice_movie/data/movie_popular/Results.kt
@@ -15,7 +15,7 @@ data class Results (
 	@SerializedName("original_title") val original_title : String,
 	@SerializedName("genre_ids") val genre_ids : List<Int>,
 	@SerializedName("title") val title : String,
-	@SerializedName("vote_average") val vote_average : Int,
+	@SerializedName("vote_average") val vote_average : Double,
 	@SerializedName("overview") val overview : String,
 	@SerializedName("release_date") val release_date : String
 )
